[fix] 스터디 메인,유저 라우트 함수 위치 이동
